ch4/ofi: Change max_buffered_send to uint16_t
The variable array allocation of a buffer for small sends is triggering
a warning about potentially using too much stack memory. This buffer is
almost certainly not going to be above UINT16_MAX in the forseable
future so cap `max_buffered_send` at that value to avoid the runaway
stack memory warning.

+++ src/mpid/ch4/netmod/ofi/ofi_types.h
@@ -347,7 +347,7 @@ typedef struct {
     struct fid_stx *stx_ctx;    /* shared TX context for RMA */
 
     /* Queryable limits */
-    uint64_t max_buffered_send;
+    uint16_t max_buffered_send;
     uint64_t max_buffered_write;
     uint64_t max_send;
     uint64_t max_write;
